requirements.txt
streamlit==1.33.0
numpy==1.24.4
tensorflow==2.13.0     # only if actually used
torch==2.0.1            # only if actually used
joblib==1.2.0
scikit-learn==1.2.2
pillow==9.4.0
